Created a 'Shared' directory and hooked up build settings/headers for it
There's some code that's used by multiple projects, including, notably,
config-file save/restore. To better manage this shared stuff, created a
"Shared" directory, and put a new source/header file pair VireioUtil in
it.
