+++ DxProxy/DxProxy/Vireio.cpp
@@ -108,24 +108,4 @@ namespace vireio {
 	{
 		*pfToClamp > max ? *pfToClamp = max : (*pfToClamp < min ? *pfToClamp = min : *pfToClamp = *pfToClamp);
 	}
-
-	std::string retprintf(const char *fmt, ...)
-	{
-		va_list args;
-		va_start(args, fmt);
-		char buf[8192];
-		vsnprintf_s(buf, 8192, fmt, args);
-		va_end(args);
-		return std::string(buf);
-	}
-	
-	void debugf(const char *fmt, ...)
-	{
-		va_list args;
-		va_start(args, fmt);
-		char buf[8192];
-		vsnprintf_s(buf, 8192, fmt, args);
-		va_end(args);
-		OutputDebugString(buf);
-	}
 };
