+++ DxProxy/DxProxy/Vireio.h
@@ -58,8 +58,5 @@ namespace vireio {
 	void UnWrapTexture(IDirect3DBaseTexture9* pWrappedTexture, IDirect3DBaseTexture9** ppActualLeftTexture, IDirect3DBaseTexture9** ppActualRightTexture);
 	bool AlmostSame(float a, float b, float epsilon);
 	void clamp(float* toClamp, float min, float max);
-
-	std::string retprintf(const char *fmt, ...);
-	void debugf(const char *fmt, ...);
 };
 #endif
