+++ VireioDLLInstaller/VireioDLLInstaller.vcxproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup Label="ProjectConfigurations">
     <ProjectConfiguration Include="Debug|Win32">
@@ -164,6 +164,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\DxProxy\DxProxy\ProxyHelper.h" />
+    <ClInclude Include="..\Shared\VireioUtil.h" />
     <ClInclude Include="..\DxProxy\DxProxy\pugiconfig.hpp" />
     <ClInclude Include="..\DxProxy\DxProxy\pugixml.hpp" />
     <ClInclude Include="Resource.h" />
@@ -171,6 +172,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\DxProxy\DxProxy\ProxyHelper.cpp" />
+    <ClCompile Include="..\Shared\VireioUtil.cpp" />
     <ClCompile Include="..\DxProxy\DxProxy\pugixml.cpp" />
     <ClCompile Include="VireioDLLInstaller.cpp" />
   </ItemGroup>
